back-end/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
beautifulsoup4==4.12.2
bokeh==3.3.2
boto3==1.34.1
botocore==1.34.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.3
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.2.0
fastapi==0.105.0
fonttools==4.46.0
frozenlist==1.4.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.1
langchain-experimental==0.0.47
langsmith==0.0.70
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.4.0
packaging==23.2
pandas==2.1.4
pdfminer==20191125
pdfminer.six==20221105
Pillow==10.1.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
s3transfer==0.9.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.2
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
validators==0.22.0
weaviate-client==3.25.3
word2number==1.1
xyzservices==2023.10.1
yarl==1.9.4
